% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_eex.R
\name{connect_eex}
\alias{connect_eex}
\title{connect_eex}
\usage{
connect_eex(path, query = NULL, root = "https://energydata.info",
  token = "")
}
\arguments{
\item{path}{character: path to send request}

\item{query}{character}

\item{root}{character: the root url that hosts the API}

\item{token}{character: access token}
}
\value{
object of class eex_api
}
\description{
connect_eex
}
