% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eex_metadata.R
\name{extract_eex_metadata}
\alias{extract_eex_metadata}
\title{extract_eex_metadata}
\usage{
extract_eex_metadata(eex_internal_id)
}
\arguments{
\item{eex_internal_id}{character: Energy Portal dataset ID}
}
\value{
list
}
\description{
Extract specific metadata from the Energy Portal API JSON response
}
