% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eex_datasets.R
\name{get_eex_datasets}
\alias{get_eex_datasets}
\title{get_eex_datasets}
\usage{
get_eex_datasets(token, origin)
}
\arguments{
\item{token}{character: Energy.Info API authentication token}

\item{origin}{character:  Flag for checking whether dataset was created in Energy.Info
can be either "True: or "False"}
}
\value{
data.frame
}
\description{
get_eex_datasets
}
