% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eex_from_ddh_datasets.R
\name{get_eex_from_ddh_datasets}
\alias{get_eex_from_ddh_datasets}
\title{get_eex_from_ddh_datasets}
\usage{
get_eex_from_ddh_datasets(root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the get_credentials() function}
}
\value{
data frame
}
\description{
Retrieve Energy info datasets on DDH
}
