% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_country_region.R
\name{map_country_region}
\alias{map_country_region}
\title{map_country_region}
\usage{
map_country_region(metadata_list, output)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}

\item{output}{list: output of map_eex_metadata_dataset()}
}
\value{
list
}
\description{
Extract specific metadata from the Finance API JSON response
}
