% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_country_values.R
\name{map_country_values}
\alias{map_country_values}
\title{map_country_values}
\usage{
map_country_values(metadata_list, output)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}

\item{output}{list: output of map_eex_metadata_dataset()}
}
\value{
list
}
\description{
Map Country values from the EEX to DDH
}
