% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_eex_metadata.R
\name{map_eex_metadata}
\alias{map_eex_metadata}
\title{map_eex_metadata}
\usage{
map_eex_metadata(metadata_list)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}
}
\value{
list
}
\description{
Extract specific metadata from the Finance API JSON response
}
