% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_eex_metadata_resource.R
\name{map_eex_metadata_resource}
\alias{map_eex_metadata_resource}
\title{map_eex_metadata_resource}
\usage{
map_eex_metadata_resource(metadata_list)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}
}
\value{
list
}
\description{
Map simple values from the EEX to DDH
}
