% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_external_metadata.R
\name{map_external_metadata}
\alias{map_external_metadata}
\title{map_external_metadata}
\usage{
map_external_metadata(metadata_list, output)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}

\item{output}{list: output of map_eex_metadata_dataset()}
}
\value{
list
}
\description{
Map External Metadata values from the EEX to DDH
}
