% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_license.R
\name{map_license}
\alias{map_license}
\title{map_license}
\usage{
map_license(metadata_list, output)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}

\item{output}{list: output of map_eex_metadata_dataset()}
}
\value{
list
}
\description{
Map License values from the EEX to DDH
}
