% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_license_values.R
\name{map_license_values}
\alias{map_license_values}
\title{map_license_values}
\usage{
map_license_values(metadata_list, output)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}

\item{output}{list: output of map_eex_metadata_dataset()}
}
\value{
list
}
\description{
Map License values from the EEX to DDH
}
