% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_multiple_values.R
\name{map_multiple_values}
\alias{map_multiple_values}
\title{map_multiple_values}
\usage{
map_multiple_values(metadata_list, output, eex_input, machine_name)
}
\arguments{
\item{metadata_list}{list: output of extract_eex_metadata()}

\item{output}{list: output of map_eex_metadata_dataset()}

\item{eex_input}{character: Energy Data API field}

\item{machine_name}{character: DDH Machine Name}
}
\value{
list
}
\description{
Map Country and Region values from the EEX to DDH
}
