% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_existing_dataset.R
\name{update_existing_dataset}
\alias{update_existing_dataset}
\title{update_existing_dataset}
\usage{
update_existing_dataset(metadata_list,
  master = eex2ddh::get_ddh_records_status(),
  ddh_fields = ddhconnect::get_fields(), lovs = ddhconnect::get_lovs(),
  root_url = dkanr::get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{metadata_list}{list: list of finance metadata, from extract_eex_metadata()}

\item{master}{dataframe: Output of fin2ddh::get_ddh_records_status()}

\item{ddh_fields}{dataframe: table of all the data catalog fields by node type}

\item{lovs}{dataframe: lookup table of the data catalog tids and values}

\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the dkanr::get_credentials() function}
}
\value{
character
}
\description{
Update a full Energy.info record in DDH (metadata + resources)
}
