% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_resources_2.R
\name{update_resources_2}
\alias{update_resources_2}
\title{update_resources_2}
\usage{
update_resources_2(dataset_nid, resource_nid, metadata_resources,
  ddh_fields = ddhconnect::get_fields(), lovs = ddhconnect::get_lovs(),
  root_url = dkanr::get_url(), credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{dataset_nid}{string: dataset to attach resource to}

\item{resource_nid}{vector: vector of resources to attach}

\item{metadata_resources}{list: list of resource metadata}

\item{lovs}{dataframe: lookup table of the data catalog tids and values}

\item{root_url}{character: API root URL}

\item{credentials}{list: object returned by the dkanr::get_credentials() function}
}
\value{
list
}
\description{
Update resources for a given dataset
}
