% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eex_datasets_list.R
\name{get_eex_datasets}
\alias{get_eex_datasets}
\title{get_eex_datasets_list}
\usage{
get_eex_datasets(token)
}
\arguments{
\item{token}{character: Energy.Info API authentication token}
}
\value{
data.frame
}
\description{
get_eex_datasets_list
}
